ISSUE #518 Refactored NWD importer into file_processor and data_processor
